rotor endstop, diskspace warning
added optional rotor endstop
fixed wrong stacking order and blurry images
added error warning if an image set is about to surpass the available diskspace
