+++ README.md
@@ -1,8 +1,18 @@
 # OpenScan2 - 3D Scanner
-## Related and more specific repositories
+
+Please take a look at the current [CHANGELOG](./docs/changelog.md)
+
+This is not the official OpenScan repository. This is a fork of [Thomas Megel](https://github.com/OpenScanEu) work but completly user-centric. Openscan is great in both software and hardware an
+and Thomas has done a great job. But openscan has those things that:
+* Users hate and complain continually and there is no one fixing nor acepting fixes/features
+* The workflow has ups and downs and some interactions feel buggy.
+The aim of this project is to tackle both weaknesses. COnvert this to a pleasant experience fro the burning of the image to the exporting of the model from the scanner.
+
+## Official Openscan repositories:
 
 If you want to take part in the development of a specific part of the OpenScan system, feel free to join:
 
+* [OpenScan2 - Official Openscan files](https://github.com/OpenScanEu/OpenScan2)
 * [OpenScanCloud - Web API for photogrammetry processing of image files](https://github.com/OpenScanEu/OpenScanCloud)
 * [OpenScan-Design - 3D printable files and other design approaches](https://github.com/OpenScanEu/OpenScan-Design)
 * [OpenScan-PCB - A place to discuss and improve the PCB designs](https://github.com/OpenScanEu/OpenScan-PCB)
@@ -10,6 +20,6 @@ If you want to take part in the development of a specific part of the OpenScan s
 
 ## Contribution and contributors
 
-The project is based on the contribution of many great and open-minded people by doing tutorials on Youtube, comments on Reddit, publications on GitHub and many other places. Without all those voluntary contributors, this project would not be possible at all. Please feel free to join the discussions and development preferably in this repository or on [r/OpenScan](https://www.reddit.com/r/OpenScan/), [Facebook - LowBudget3DScan](https://www.facebook.com/groups/142108429832711) or [OpenScan.eu/forum](https://openscan.eu/forum)
+The official project is based on the contribution of many great and open-minded people by doing tutorials on Youtube, comments on Reddit, publications on GitHub and many other places. Without all those voluntary contributors, this project would not be possible at all. Please feel free to join the discussions and development preferably in this repository or on [r/OpenScan](https://www.reddit.com/r/OpenScan/), [Facebook - LowBudget3DScan](https://www.facebook.com/groups/142108429832711) or [OpenScan.eu/forum](https://openscan.eu/forum)
 
 Thank you!
