+++ update/beta/OpenScan.py
@@ -1,5 +1,6 @@
 basepath = '/home/pi/OpenScan/'
 from os.path import isfile
+import os
 
 def load_bool(name):
     filename = basepath+'settings/'+name
@@ -12,6 +13,74 @@ def load_bool(name):
     else:
         value = False
     return value
+
+def fade_led(pin_led, fade_steps, duty_max, dir = True):
+    import RPi.GPIO as GPIO
+    import time
+    GPIO.setmode(GPIO.BCM)
+    GPIO.setwarnings(False)
+    GPIO.setup(pin_led, GPIO.OUT)
+    pwm = GPIO.PWM(pin_led, 200)
+
+    if dir:
+        pwm.start(0)
+        for duty_cycle in range(0, fade_steps*10, 1):  # Increase duty cycle in steps
+            pwm.ChangeDutyCycle(duty_max*duty_cycle/(10*fade_steps))
+            time.sleep(0.001)  # Pause between steps (adjust as needed)
+    else:
+        pwm.start(duty_max)
+        for duty_cycle in range(fade_steps*10,0, -1):  # Increase duty cycle in steps
+            pwm.ChangeDutyCycle(duty_max*duty_cycle/(10*fade_steps))
+            time.sleep(0.001)  # Pause between steps (adjust as needed)
+    pwm.stop()
+
+
+def check_hotspot_mode(interface="wlan0"):
+    import subprocess
+    try:
+        output = subprocess.check_output(["iwconfig", interface]).decode("utf-8")
+        if "Mode:Master" in output:
+            return True
+        elif "Mode:Managed" in output:
+            return False
+        else:
+            return False
+    except subprocess.CalledProcessError as e:
+        return False
+
+
+
+def add_wifi_network(ssid, password, country):
+    import re
+    conf_file = "/etc/wpa_supplicant/wpa_supplicant-wlan0.conf"
+
+    if not os.path.exists(conf_file):
+        return False
+
+    if not (ssid and password and country):
+        return False
+
+    with open(conf_file, "r") as f:
+        content = f.read()
+
+    updated_content = re.sub(r'country=\w+', f'country={country}', content)
+
+    if f'ssid="{ssid}"' in content:
+        network_block_pattern = re.compile(
+            r'network=\{\s*ssid="' + re.escape(ssid) + r'".*?psk=".*?".*?\}', re.DOTALL
+        )
+        updated_network_block = f'network={{\n    ssid="{ssid}"\n    psk="{password}"\n    key_mgmt=WPA-PSK\n}}'
+        updated_content = network_block_pattern.sub(updated_network_block, updated_content)
+    else:
+        network_block = f'\nnetwork={{\n    ssid="{ssid}"\n    psk="{password}"\n    key_mgmt=WPA-PSK\n}}\n'
+        updated_content += network_block
+
+    with open(conf_file, "w") as f:
+        f.write(updated_content)
+    os.system("sudo systemctl restart wpa_supplicant@wlan0")
+
+    return True
+
 
 def load_str(name):
     filename = basepath+'settings/'+name
@@ -70,7 +139,6 @@ def motorrun(motor,angle,ES_enable=False,ES_start_state = True):
     from time import sleep
     from math import cos
     msg = {'cmd':'set'}
-    camera('/ping', msg)
 
     GPIO.setwarnings(False)
     GPIO.setmode(GPIO.BCM)
@@ -97,7 +165,14 @@ def motorrun(motor,angle,ES_enable=False,ES_start_state = True):
         step_count=-step_count
     for x in range(step_count):
         if ES_enable == True and GPIO.input(ES_pin) != ES_start_state:
-            break
+            i = 0
+            while i <= 10:
+                if GPIO.input(ES_pin) == ES_start_state:
+                    i = 11
+                if i == 10:
+                    return
+                i = i + 1
+
         GPIO.output(steppin, GPIO.HIGH)
         if x<=ramp and x<=step_count/2:
             delay = delay_init * (1 + -1/acc*cos(1*(ramp-x)/ramp)+1/acc)
@@ -114,7 +189,6 @@ def motorrun(motor,angle,ES_enable=False,ES_start_state = True):
 def ringlight(number,state):
     import RPi.GPIO as GPIO
     msg = {'cmd':'set'}
-    camera('/ping', msg)
     pin = load_int('pin_ringlight' + str(number))
     GPIO.setwarnings(False)
     GPIO.setmode(GPIO.BCM)
@@ -201,3 +275,42 @@ def create_coordinates(angle_min, angle_max,point_count):
             point_count=point_count+1
     return filtered
 
+
+def haversine_distance_deg(theta1, phi1, theta2, phi2):
+    import numpy as np
+    R = 1
+    dtheta = np.radians(theta2 - theta1)
+    dphi = np.radians(phi2 - phi1)
+
+    theta1, phi1 = np.radians(theta1), np.radians(phi1)
+    theta2, phi2 = np.radians(theta2), np.radians(phi2)
+
+    a = np.sin(dtheta / 2) ** 2 + np.cos(theta1) * np.cos(theta2) * np.sin(dphi / 2) ** 2
+    c = 2 * np.arctan2(np.sqrt(a), np.sqrt(1 - a))
+
+    return R * c
+
+def sort_spherical_coordinates_deg(points_spherical_deg):
+    import numpy as np
+    from tsp_solver.greedy import solve_tsp
+
+    points_spherical_deg = np.array(points_spherical_deg)  # Convert list of tuples to NumPy array
+
+    n = len(points_spherical_deg)
+    dist_matrix = np.zeros((n, n))
+
+    # Calculate haversine distance for each pair of points
+    for i in range(n):
+        for j in range(i + 1, n):
+            dist = haversine_distance_deg(points_spherical_deg[i, 0], points_spherical_deg[i, 1],
+                                          points_spherical_deg[j, 0], points_spherical_deg[j, 1])
+            dist_matrix[i, j] = dist
+            dist_matrix[j, i] = dist
+
+    # Solve the TSP problem using the tsp_solver.greedy algorithm
+    path = solve_tsp(dist_matrix)
+
+    sorted_points_spherical_deg = points_spherical_deg[path]
+
+    # Convert the sorted NumPy array back to a list of tuples
+    return [tuple(point) for point in sorted_points_spherical_deg]
